requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0